Create Node class and tests
I created a node class to represent a single field for the connect4 game
which has it's place on the field stored as it's value as well as an
instance variable to store the player's symbol.
It also stores each neighbor as an instance variable named after the
location on the field in relation to the node itself, which I will
utilize to check the winning condition of the game.