Rename field to board and create class
I created a field class which upon initialization creates the board used
for the game
Without further input the board consists of 42 Node object to represent
the typical connect4 board with 7 rows and 6 lines.
The class also has a method to display the board as well as the
corresponding number for each line for the player to choose where to
place the token.