Create individual check for four connected symbols
This contains the first draft of methods to check whether there are 4
similar symbols on a lane.
Each of the 4 possible lane-types has its own method to count the number
of similar symbols in a lane.
For the final game, I replaced these 4 methods with a single one, which
finds the starting node in each direction and also returns the number of
connecting similar symbols in a lane.