Create Connect4 class with game logic
I created a new class Connect4 which contains all the methods necessary
to play the game.
Everything was created TDD style and the game passes all the tests.
The class has too many lines according to rubocop and I am sure there is
a lot of room to improve several methods, but as the game works,
I move refactoring further into the future until I might have learned
some new tricks. Then I will also include an option for the player to
choose the size of the board.